requirements.txt
asgiref==3.5.2
attrs==22.2.0
beautifulsoup4==4.11.2
Django==3.2.16
django-bootstrap5==22.2
django-debug-toolbar==4.2.0
Faker==12.0.1
flake8==5.0.4
iniconfig==2.0.0
isort==5.12.0
mccabe==0.7.0
mixer==7.2.2
packaging==23.0
Pillow==9.3.0
pluggy==1.0.0
py==1.11.0
pycodestyle==2.9.1
pyflakes==2.5.0
pytest==7.1.3
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2022.7
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.3
tomli==2.0.1
yapf==0.32.0
